VMS updates. Thanks Bruce Visscher.
git-svn-id: https://omniorb.svn.sourceforge.net/svnroot/omniorb/branches/4_1@6046 31ce16a7-c309-44a6-b7c4-d857baf782c7

+++ omniORB/src/lib/omniORB/orbcore/corbaFixed.cc
@@ -177,7 +177,7 @@ CORBA::Fixed::Fixed(CORBA::Double val) :
 		  CORBA::COMPLETED_NO);
   }
   char buffer[80];
-  int len = sprintf(buffer, "%.31f", val);
+  int len = sprintf(buffer, "%.31f", (double)val);
   OMNIORB_ASSERT(len < 79);
 
   NP_fromString(buffer);
